feat(contacts): contacts provider done

+++ src/providers/emails-providers/emails-providers.controller.ts
@@ -13,6 +13,8 @@ import {
   Inject,
   Put,
   NotFoundException,
+  Request,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { CreateEmailsProviderDto } from './dto/create-emails-provider.dto';
 import { UpdateEmailsProviderDto } from './dto/update-emails-provider.dto';
@@ -28,6 +30,9 @@ import { EmailsProvider } from './schemas/emails-provider.schema';
 import { EmailsProvidersService } from './emails-providers.service';
 import { Domain } from './schemas/domain.schema';
 import { UpdateSendersDto } from './dto/update-senders.dto';
+import { AuthenticatedRequest } from '../../shared/types/authentication.types';
+import { Action } from '../../rbac/permissions/entities/action.entity';
+import { ScopedSubject } from '../../rbac/casl/casl.utils';
 
 @Controller('organizations/:organizationId/providers/emails')
 @UseGuards(OrganizationGuard)
@@ -53,6 +58,7 @@ export class EmailsProvidersController {
 
   @Post()
   async create(
+    @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Body() createEmailsProviderDto: CreateEmailsProviderDto,
   ): Promise<EmailsProvider> {
@@ -61,6 +67,19 @@ export class EmailsProvidersController {
         'Organization ID does not match provider organization ID',
       );
     }
+
+    if (
+      req.ability.cannot(
+        Action.Create,
+        ScopedSubject(EmailsProvider, organizationId),
+      )
+    ) {
+      throw new UnauthorizedException('Cannot create emails provider');
+    }
+
+    this.logger.log(
+      'Creating emails provider for organization ' + organizationId,
+    );
 
     const providers = await this.emailsProvidersService.findAll(organizationId);
 
@@ -76,8 +95,18 @@ export class EmailsProvidersController {
 
   @Get()
   findAll(
+    @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
   ): Promise<EmailsProvider[]> {
+    if (
+      req.ability.cannot(
+        Action.Read,
+        ScopedSubject(EmailsProvider, organizationId),
+      )
+    ) {
+      throw new UnauthorizedException('Cannot read emails providers');
+    }
+
     return this.emailsProvidersService.findAll(organizationId);
   }
 
@@ -85,9 +114,19 @@ export class EmailsProvidersController {
   @CacheTTL(300000)
   @UseInterceptors(CacheInterceptor)
   async findOne(
+    @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('providerId') providerId: string,
   ): Promise<EmailsProvider> {
+    if (
+      req.ability.cannot(
+        Action.Read,
+        ScopedSubject(EmailsProvider, organizationId),
+      )
+    ) {
+      throw new UnauthorizedException('Cannot read emails provider');
+    }
+
     const providerDocument = await this.emailsProvidersService.findOne(
       organizationId,
       providerId,
@@ -102,10 +141,24 @@ export class EmailsProvidersController {
 
   @Patch(':providerId')
   async update(
+    @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('providerId') providerId: string,
     @Body() updateEmailsProviderDto: UpdateEmailsProviderDto,
   ): Promise<EmailsProvider> {
+    if (
+      req.ability.cannot(
+        Action.Update,
+        ScopedSubject(EmailsProvider, organizationId),
+      )
+    ) {
+      throw new UnauthorizedException('Cannot update emails provider');
+    }
+
+    this.logger.log(
+      'Updating emails provider of organization ' + organizationId,
+    );
+
     const needToSetDefault = updateEmailsProviderDto.isDefault;
 
     let previousDefaultProvider: EmailsProvider;
@@ -139,9 +192,23 @@ export class EmailsProvidersController {
 
   @Delete(':providerId')
   async remove(
+    @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('providerId') providerId: string,
   ) {
+    if (
+      req.ability.cannot(
+        Action.Delete,
+        ScopedSubject(EmailsProvider, organizationId),
+      )
+    ) {
+      throw new UnauthorizedException('Cannot delete emails provider');
+    }
+
+    this.logger.log(
+      'Removing emails provider of organization ' + organizationId,
+    );
+
     const deletion = await this.emailsProvidersService.remove(
       organizationId,
       providerId,
@@ -156,10 +223,22 @@ export class EmailsProvidersController {
 
   @Put(':providerId/senders')
   async updateSenders(
+    @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('providerId') providerId: string,
     @Body('senders') senders: UpdateSendersDto,
   ): Promise<EmailsProvider> {
+    if (
+      req.ability.cannot(
+        Action.Update,
+        ScopedSubject(EmailsProvider, organizationId),
+      )
+    ) {
+      throw new UnauthorizedException(
+        'Cannot update senders of emails provider',
+      );
+    }
+
     this.logger.log(
       `Updating ${senders.length} senders for provider ${providerId}`,
     );
@@ -186,10 +265,22 @@ export class EmailsProvidersController {
 
   @Post(':providerId/domains')
   async createDomain(
+    @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('providerId') providerId: string,
     @Body('domain') domain: string,
   ): Promise<void> {
+    if (
+      req.ability.cannot(
+        Action.Update,
+        ScopedSubject(EmailsProvider, organizationId),
+      )
+    ) {
+      throw new UnauthorizedException(
+        'Cannot create domain for emails provider',
+      );
+    }
+
     this.logger.log(`Creating domain ${domain} for provider ${providerId}`);
 
     const providerDocument = await this.emailsProvidersService.findOne(
@@ -205,10 +296,22 @@ export class EmailsProvidersController {
 
   @Delete(':providerId/domains/:domain')
   async deleteDomain(
+    @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('providerId') providerId: string,
     @Param('domain') domain: string,
   ): Promise<void> {
+    if (
+      req.ability.cannot(
+        Action.Update,
+        ScopedSubject(EmailsProvider, organizationId),
+      )
+    ) {
+      throw new UnauthorizedException(
+        'Cannot delete domain of emails provider',
+      );
+    }
+
     this.logger.log(`Deleting domain ${domain} of provider ${providerId}`);
 
     const providerDocument = await this.emailsProvidersService.findOne(
@@ -224,10 +327,20 @@ export class EmailsProvidersController {
 
   @Post(':providerId/domains/:domain/dns')
   async checkDomain(
+    @Request() req: AuthenticatedRequest,
     @Param('organizationId') organizationId: string,
     @Param('providerId') providerId: string,
     @Param('domain') domain: string,
   ): Promise<Domain> {
+    if (
+      req.ability.cannot(
+        Action.Update,
+        ScopedSubject(EmailsProvider, organizationId),
+      )
+    ) {
+      throw new UnauthorizedException('Cannot check domain of emails provider');
+    }
+
     this.logger.log(`Checking domain ${domain} for provider ${providerId}`);
 
     const providerDocument = await this.emailsProvidersService.findOne(
